append.c: use long long integers for %ll... format
Since we *may* need to use 64-bit arithmetic on machines where long is
32-bit (Windows), switch handle_decimal, append_long and append_ulong
to long long arguments.

Add tests to make sure that ULL constants are correctly formatted
as signed decimal / unsigned decimal / hexadecimal (which should be
unsigned). Ignore the octal, which for now is signed.

+++ append.h
@@ -45,7 +45,7 @@ char	*append_string(char *dest, const char *limit, const char *value);
  * Chmela which is released under GPLv3.
  */
 extern
-char	*append_long(char *dest, char *limit, long value, int base);
+char	*append_long(char *dest, char *limit, long long value, int base);
 
 /*
  * Append unsigned long value argument to destination up to limit
@@ -54,7 +54,7 @@ char	*append_long(char *dest, char *limit, long value, int base);
  * written by Lukas Chmela. Released under GPLv3.
  */
 extern
-char	*append_ulong(char *dest, char *limit, unsigned long value, int base);
+char	*append_ulong(char *dest, char *limit, unsigned long long value, int base);
 
 /*
  * Append pointer value argument to destination up to limit pointer.
